+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - fix: do not call flags callback with invalid flags ([#355](https://github.com/PostHog/posthog-ios/pull/355))
+- use new `/flags` endpoint instead of `/decide` ([#345](https://github.com/PostHog/posthog-ios/pull/345))
 
 ## 3.26.2 - 2025-06-03